# clinical-mdr-api/consumer_api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "StudyBuilder Consumer API",
...
    "version": "0.1.90"
  },
  "paths": {
...
    "/system/information": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Returns various information about this API (running version, etc.)",
        "operationId": "get_system_information_system_information_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemInformation"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/information/build-id": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Returns build id as plain text",
        "operationId": "get_build_id_system_information_build_id_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Build Id System Information Build Id Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/information/sbom.md": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Returns SBOM as markdown text",
        "operationId": "get_sbom_md_system_information_sbom_md_get",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/information/license.md": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Returns license as markdown text",
        "operationId": "get_license_md_system_information_license_md_get",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Exception class."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "More information about the error."
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time",
            "description": "The point in time when the error occurred."
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "The url/path of the request."
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "The HTTP method of the request."
          }
        },
        "type": "object",
        "required": [
          "type",
          "message",
          "time",
          "path",
          "method"
        ],
        "title": "ErrorResponse"
      },
...
      "SystemInformation": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "description": "Version of the API specification"
          },
          "db_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Db Version",
            "description": "Version information from the Neo4j database the application is using"
          },
          "db_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Db Name",
            "description": "Name of the database the application is using"
          },
          "build_id": {
            "type": "string",
            "title": "Build Id",
            "description": "The Build.BuildNumber identifier from the pipeline run"
          },
          "commit_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Id",
            "description": "The reference to the repository state: the id of the last commit to the branch at build"
          },
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name",
            "description": "Name of the VCS repository branch the app was built from"
          }
        },
        "type": "object",
        "required": [
          "api_version",
          "build_id"
        ],
        "title": "SystemInformation"
      },
...
    },
    "securitySchemes": {
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "api:///API.call": "Make calls to the API"
            },
            "authorizationUrl": "",
            "tokenUrl": ""
          }
        }
      },
      "BearerJwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "description": "Access token that will be sent as `Authorization: Bearer {token}` header in all requests"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}